Update credit card name if billing names updated
This wasn't creating the card name for me: `card_name += $(el).val() + ' ';`. I was getting a space, but now first+last name. (Perhaps because of a woocommerce modification in the installation I'm working with).

The rest of the code I added is probably a little ick, but my intentions are good; always pass the current first+last name to Stripe.